+++ changelog.md
@@ -1,8 +1,16 @@
+Version 1.2.0
+-------------
+* mouse: fix `clicks` not works in wheelUp and wheelDown.
+* process: add getCurrentProcess, getProcess, getCommandLine, getStats,
+  suspend, resume, stdioClose, stdinWrite, stdoutRead, stderrRead,
+  run, runAs, runWait, runAsWait, shellExecute, shellExecuteWait.
+* add misc submodule, provide isAdmin and requireAdmin.
+
 Version 1.1.0
 -------------
 * window: add minimizeAll, minimizeAllUndo.
 * process: add isWow64 (#2).
-* registry: Fix RegMultiSz writing with regWrite (#5).
+* registry: fix RegMultiSz writing with regWrite (#5).
 
 Version 1.0.0
 -------------